+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixes a bug where OData select query parameter would not be normalized
 - Fixes a bug in Go where empty collections would not be serialized.
 - Fixes a bug where Java generation would fail because of empty usings.
+- Fixes a bug where Java and Go escaped model properties would not serialize properly.
 
 ## [0.0.14] - 2021-11-08